Move EnvironmentExtensions.cs into the Extensions folder
(cherry picked from commit a8602023ca6a472fcd0a26aa5eb215b0221cf1ff)
