+++ DfE.FindInformationAcademiesTrusts/Setup/SecurityServicesSetup.cs
@@ -2,6 +2,7 @@
 using System.Diagnostics.CodeAnalysis;
 using Azure.Identity;
 using DfE.FindInformationAcademiesTrusts.Authorization;
+using DfE.FindInformationAcademiesTrusts.Extensions;
 using Microsoft.AspNetCore.Authentication.Cookies;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.DataProtection;
