+++ tests/DfE.FindInformationAcademiesTrusts.UnitTests/Extensions/EnvironmentExtensionsTests.cs
@@ -1,6 +1,7 @@
+using DfE.FindInformationAcademiesTrusts.Extensions;
 using Microsoft.AspNetCore.Hosting;
 
-namespace DfE.FindInformationAcademiesTrusts.UnitTests;
+namespace DfE.FindInformationAcademiesTrusts.UnitTests.Extensions;
 
 public class EnvironmentExtensionsTests
 {
